Move DimensionMetrics to own class, add Builder
